refactor: split events controller into stream + interrupts (#2253)
